Output polex based on polexdraft
